feat: 닉네임 중복 확인 API 및 관련 서비스 로직 제거 (#409)
